requirements.txt
flask
flask-cors
numpy
pandas
joblib
scikit-learn
tensorflow-cpu==2.20.0
gunicorn
keras>=3.4.0